better color mapping of term colors to neovim highlights
we use wallust to generate colors. it creates a palette from desktop
background and orders by luminance. there is no reliable hue due to
changing desktops. we cannot use hue as a means to signify attention. we
can, however, use luminance to do this.

colors are mapped on a imperative model.
verb > noun > context

the more important, the more luminance it should be given.